@@ -1,6 +1,6 @@
 <?php
 /**
- *  Copyright (C) 2010 - 2020  <Robbyn Gerhardt>
+ *  Copyright (C) 2010 - 2021  <Robbyn Gerhardt>
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -17,7 +17,7 @@
  *
  * @package       webpackages
  * @author        Robbyn Gerhardt
- * @copyright     Copyright (c) 2010 - 2020
+ * @copyright     Copyright (c) 2010 - 2021
  * @license       http://opensource.org/licenses/MIT	MIT License
  * @since         Version 2.0.0
  * @filesource
@@ -25,6 +25,7 @@
 
 namespace system\core\Logger;
 
+use system\core\Logger\Adapter\AdapterInterface;
 use system\core\Logger\Adapter\filelogger;
 
 class logger
@@ -32,9 +33,9 @@ class logger
 	/**
 	 * @param LoggerConfig $config
 	 *
-	 * @return filelogger
+	 * @return AdapterInterface
 	 */
-	public static function create(LoggerConfig $config)
+	public static function create(LoggerConfig $config): AdapterInterface
 	{
 		if($config->engine == 'filelogger')
 		{
