@@ -1,6 +1,6 @@
 <?php
 /**
- *  Copyright (C) 2010 - 2020  <Robbyn Gerhardt>
+ *  Copyright (C) 2010 - 2021  <Robbyn Gerhardt>
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -17,7 +17,7 @@
  *
  * @package       webpackages
  * @author        Robbyn Gerhardt
- * @copyright     Copyright (c) 2010 - 2020
+ * @copyright     Copyright (c) 2010 - 2021
  * @license       http://opensource.org/licenses/MIT	MIT License
  * @since         Version 2.0.0
  * @filesource
@@ -25,11 +25,14 @@
 
 namespace system\core;
 
+use FilesystemIterator;
+use RecursiveDirectoryIterator;
+use SplFileInfo;
 use system\plugins\Adapter\AdapterPlugins;
 
 class Plugin
 {
-	private static $_hooks	=	[];
+	private static array $_hooks	=	[];
 
 	/**
 	 * Call hook
@@ -37,7 +40,7 @@ class Plugin
 	 * @param string $name
 	 * @param null $params
 	 */
-	public static function hook($name, $params = null)
+	public static function hook(string $name, $params = null)
 	{
 		if(isset(self::$_hooks[$name]))
 		{
@@ -54,7 +57,7 @@ class Plugin
 	 * @param string $hook
 	 * @param array $call
 	 */
-	public static function register($hook, array $call)
+	public static function register(string $hook, array $call)
 	{
 		self::$_hooks[$hook][]	=	$call;
 	}
@@ -65,10 +68,10 @@ class Plugin
 	public static function loadPluginRegister()
 	{
 		$path	=	ROOT.'system'.DIRECTORY_SEPARATOR.'plugins'.DIRECTORY_SEPARATOR;
-		$dirs	=	new \RecursiveDirectoryIterator($path, \RecursiveDirectoryIterator::SKIP_DOTS);
+		$dirs	=	new RecursiveDirectoryIterator($path, FilesystemIterator::SKIP_DOTS);
 
 		/**
-		 * @var \SplFileInfo $dir
+		 * @var SplFileInfo $dir
 		 */
 		foreach($dirs as $dir)
 		{
